Refactor: 상품 목록 페이지 로그인 세션을 포함한 일부 로직 및 메서드, 변수명 변경
