+++ src/main/java/com/needle/FsoFso/product/service/ProductService.java
@@ -1,0 +1,14 @@
+package com.needle.FsoFso.product.service;
+
+import java.util.List;
+
+import com.needle.FsoFso.product.dto.ProductDto;
+
+public interface ProductService {
+	
+	List<ProductDto> productList();
+	
+	List<ProductDto> getproducPagelist(int start);
+	
+	int getAllProductCount();
+}
