--- src/main/java/com/needle/FsoFso/product/services/ProductService.java
@@ -1,14 +1,0 @@
-package com.needle.FsoFso.product.services;
-
-import java.util.List;
-
-import com.needle.FsoFso.product.dto.ProductDto;
-
-public interface ProductService {
-	
-	List<ProductDto> productList();
-	
-	List<ProductDto> getproducPagelist(int start);
-	
-	int getAllProduct();
-}
